Refactor "enable_when_ready" into utility module
